% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Z.R
\name{backward_algorithm}
\alias{backward_algorithm}
\title{Backward algorithm}
\usage{
backward_algorithm(e0, e1, colo, ex)
}
\arguments{
\item{e0}{the emission probabilities given non-occupancy}

\item{e1}{the emission probabilities given occupancy}

\item{colo}{a vector giving the colonization probs}

\item{ex}{a vector giving the extinction probs}
}
\value{
an n x 2 matrix giving the backward probabilities
}
\description{
Backward algorithm
}
