% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colex_init_types}
\alias{colex_init_types}
\title{possible values for the `colex_init` parameter if not NULL}
\usage{
colex_init_types()
}
\value{
character vector of options
}
\description{
possible values for the `colex_init` parameter if not NULL
}
