% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emission_likelihood}
\alias{emission_likelihood}
\title{Get emission likelihoods given observations and detection probabilities}
\usage{
emission_likelihood(state, obs, det)
}
\arguments{
\item{state}{Compute the emission likelihood for absence (0) or presence (1)}

\item{obs}{A matrix of observation histories. Rows are units, columns are visits.}

\item{det}{A matrix of detection probabilities}
}
\value{
a vector of emission likelihoods (one per row)
}
\description{
Get emission likelihoods given observations and detection probabilities
}
