% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_flocker_data}
\alias{example_flocker_data}
\title{Example flocker data}
\format{
## `example_flocker_data`
A list with elements appropriate to pass to make_flocker_data:
\describe{
  \item{obs}{A detection/nondetection matrix. Rows are units and columns are
  visits}
  \item{unit_covs}{A data.frame of four unit covs. `uc1` and `uc2` are 
  numeric; `grp` and `species` are categorical. Every level of `species` is
  present for every value of `uc1` and `uc2`, as would be the case in a 
  typical multispecies model}
  \item{event_covs}{A named list of two matrices, each representing a numeric
  covariate that varies by visit but not by species}
}
}
\source{
data-raw/example_flocker_data.R
}
\usage{
example_flocker_data
}
\description{
Example data to be formatted for use in a single-season occupancy model
}
\keyword{datasets}
