% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_flocker_model_aug}
\alias{example_flocker_model_aug}
\title{Example data-augmented flocker model}
\format{
## `example_flocker_model_aug`
A flocker_fit and brmsfit object
}
\source{
data-raw/example_flocker_model.R
}
\usage{
example_flocker_model_aug
}
\description{
A fitted data-augmented occupancy model from flocker
}
\keyword{datasets}
