% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_flocker_model_multi}
\alias{example_flocker_model_multi}
\title{Example multi-season flocker model}
\format{
## `example_flocker_model_multi`
A flocker_fit and brmsfit object
}
\source{
data-raw/example_flocker_model.R
}
\usage{
example_flocker_model_multi
}
\description{
A fitted multi-season occupancy model from flocker.
One species 200 point, four season, four rep, missing seasons, missing reps.
colex parameterization with explicit init.
}
\keyword{datasets}
