% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_flocker_model_multi_colex_eq}
\alias{example_flocker_model_multi_colex_eq}
\title{Example multi-season flocker model (colex equilibrium)}
\format{
## `example_flocker_model_multi_colex_eq`
A flocker_fit and brmsfit object
}
\source{
data-raw/example_flocker_model.R
}
\usage{
example_flocker_model_multi_colex_eq
}
\description{
A fitted multi-season occupancy model from flocker.
One species 200 point, six season, four rep, missing seasons, missing reps.
colex parameterization with equilibrium init.
}
\keyword{datasets}
