% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_flocker_model_single_C}
\alias{example_flocker_model_single_C}
\title{Example single-season rep-constant flocker model}
\format{
## `example_flocker_model_single_C`
A flocker_fit and brmsfit object
}
\source{
data-raw/example_flocker_model.R
}
\usage{
example_flocker_model_single_C
}
\description{
A fitted single-season rep-constant occupancy model from flocker
}
\keyword{datasets}
