% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_flocker_model_single_fp}
\alias{example_flocker_model_single_fp}
\title{Example single-season false-positive flocker model}
\format{
## `example_flocker_model_single_fp`
A flocker_fit and brmsfit object
}
\source{
data-raw/example_flocker_model.R
}
\usage{
example_flocker_model_single_fp
}
\description{
A fitted single-season false-positive occupancy model from flocker
}
\keyword{datasets}
