% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_array_3D}
\alias{expand_array_3D}
\title{convert array to long matrix format}
\usage{
expand_array_3D(a)
}
\arguments{
\item{a}{3D array to expand}
}
\value{
matrix
}
\description{
convert array to long matrix format
}
