% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_matrix}
\alias{expand_matrix}
\title{convert matrix to long vector format}
\usage{
expand_matrix(m)
}
\arguments{
\item{m}{matrix-like object to expand}
}
\value{
vector
}
\description{
convert matrix to long vector format
}
