% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extended_binomial_rng}
\alias{extended_binomial_rng}
\title{extended binomial RNG}
\usage{
extended_binomial_rng(p, s)
}
\arguments{
\item{p}{binomial probability}

\item{s}{number of successes}
}
\value{
a sample representing a possible number of trials
}
\description{
extended binomial RNG
}
\details{
random samples `y` from the distribution proportional to 
  binomial_pmf(y | n, s)
}
