% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fdtl}
\alias{fdtl}
\title{flocker data type lookup}
\usage{
fdtl()
}
\value{
dataframe giving lookup table for data input types, data output types
  and model types
}
\description{
flocker data type lookup
}
