% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flock.R
\name{flock_}
\alias{flock_}
\title{Fit an occupancy model or generate Stan code for a model}
\usage{
flock_(
  stancode,
  f_occ,
  f_det,
  flocker_data,
  data2 = NULL,
  multiseason = NULL,
  f_col = NULL,
  f_ex = NULL,
  colex_init = NULL,
  augmented = FALSE,
  fp = FALSE,
  ...
)
}
\arguments{
\item{stancode}{logical: output Stan code or fitted model (T) or fit model (F)}

\item{f_occ}{A brms-type model formula for occupancy. If provided, must begin 
with "~".}

\item{f_det}{A brms-type model formula for detection. Must begin with "~".}

\item{flocker_data}{data, generally the output of \code{make_flocker_data()}.}

\item{data2}{additional data (e.g. a covariance matrix for a phylogenetic 
effect)}

\item{multiseason}{Must be NULL (the default) or one of "colex" or 
"autologistic". If NULL, data must be formatted for a single-season model.
Otherwise, the data must be formatted for a multiseason model. If "colex", a 
colonization-extinction model will be fit, and `f_col` and `f_ex` must be 
specified. If "autologistic", an autologistic model will be fit, and `f_col`
and `f_ex` must both be NULL.}

\item{f_col}{A brms-type model formula for colonization in 
colonization-extinction dynamic models. If provided, must begin with "~".}

\item{f_ex}{A brms-type model formula for extinction probabilities in 
colonization-extinction dynamic models. If provided, must begin with "~".}

\item{colex_init}{Must be NULL unless the model is a colonization-extinction
dynamic model, in which case must be either "explicit" or "equilibrium". 
If "explicit", then `f_occ` must be provided to model 
of year-1 occupancy probabilities. If"equilibrium", then `f_occ` must be set
to `NULL` and the initial occupancy probabilities are assumed to 
be the (possibly site-specific) equilibrium probabilities from the 
colonization-extinction dynamics. In this case, time-varying covariates are 
not allowed for colonization or extinction (but are still allowed for 
detection).}

\item{augmented}{Logical. Must be TRUE if data are formatted for a 
data-augmented multi-species model, and FALSE otherwise.}

\item{fp}{logical. If data are formatted for an fp model, must be set to
TRUE, otherwise to FALSE.}

\item{...}{additional arguments passed to \code{brms::brm()}}
}
\value{
stan code for brms model or a \code{brmsfit} containing the fitted occupancy model.
}
\description{
Fit an occupancy model or generate Stan code for a model
}
