% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flocker_col_names}
\alias{flocker_col_names}
\title{column names created in flocker}
\usage{
flocker_col_names(n_rep = NULL, n_year = NULL)
}
\arguments{
\item{n_rep}{max number of repeat visits}

\item{n_year}{max length of hmm series in dynamic models}
}
\value{
character vector of column names created internally by 
   make_flocker_data
}
\description{
column names created in flocker
}
