% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flocker_data_input_types}
\alias{flocker_data_input_types}
\title{Options for the \code{type} argument to \code{make_flocker_data}.}
\usage{
flocker_data_input_types()
}
\value{
character vector of possible inputs
}
\description{
Options for the \code{type} argument to \code{make_flocker_data}.
}
