% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flocker_data_output_types}
\alias{flocker_data_output_types}
\title{Data types that can be returned by \code{make_flocker_data}.}
\usage{
flocker_data_output_types()
}
\value{
character vector of data output types
}
\description{
Data types that can be returned by \code{make_flocker_data}.
}
