% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flocker_make_standata.R
\name{flocker_make_standata}
\alias{flocker_make_standata}
\title{Make Stan data for a flocker model with visit-specific detection covariates.
Primarily for internal use in `flocker()`}
\usage{
flocker_make_standata(f_occ_use, f_det_use, flocker_data, data2)
}
\arguments{
\item{f_occ_use}{Occupancy formula}

\item{f_det_use}{Detection formula}

\item{flocker_data}{data, generally the output of `make_flocker_data()`.}

\item{data2}{additional data (e.g. a covariance matrix for a phylogenetic effect)}
}
\value{
A standata list.
}
\description{
Make Stan data for a flocker model with visit-specific detection covariates.
Primarily for internal use in `flocker()`
}
