% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flocker_model_types}
\alias{flocker_model_types}
\title{Model types in flocker as outputted by `flock`}
\usage{
flocker_model_types()
}
\value{
character vector of types
}
\description{
Model types in flocker as outputted by `flock`
}
