% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flocker_reserved}
\alias{flocker_reserved}
\title{Reserved names in flocker}
\usage{
flocker_reserved()
}
\value{
character vector of regexes matching reserved variable names
}
\description{
Reserved names in flocker
}
