% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formula_error}
\alias{formula_error}
\title{Create error message for formula with incorrect syntax}
\usage{
formula_error(form)
}
\arguments{
\item{form}{string giving name of formula}
}
\value{
string giving text of error
}
\description{
Create error message for formula with incorrect syntax
}
