% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Z.R
\name{forward_backward_algorithm}
\alias{forward_backward_algorithm}
\title{forward backward algorithm}
\usage{
forward_backward_algorithm(el0, el1, init, colo, ex)
}
\arguments{
\item{el0}{the emission probabilities given non-occupancy}

\item{el1}{the emission probabilities given occupancy}

\item{init}{the initial occupancy probability}

\item{colo}{a vector giving the colonization probs}

\item{ex}{a vector giving the extinction probs}
}
\value{
a vector of posterior Z probabilities
}
\description{
forward backward algorithm
}
