% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Z.R
\name{forward_backward_sampling}
\alias{forward_backward_sampling}
\title{Forward filtering backward sampling algorithm}
\usage{
forward_backward_sampling(el0, el1, init, colo, ex)
}
\arguments{
\item{el0}{the emission probabilities given non-occupancy}

\item{el1}{the emission probabilities given occupancy}

\item{init}{the initial occupancy probability}

\item{colo}{a vector giving the colonization probs}

\item{ex}{a vector giving the extinction probs}
}
\description{
Forward filtering backward sampling algorithm
}
