% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Z.R
\name{forward_sim}
\alias{forward_sim}
\title{Forward simulation for a dynamic model}
\usage{
forward_sim(init, colo, ex, sample = FALSE)
}
\arguments{
\item{init}{initial occupancy probability}

\item{colo}{vector of colonization probabilities}

\item{ex}{vector of extinction probabilities}

\item{sample}{Should the return be posterior probabilities of occupancy (FALSE),
valid posterior predictive samples (TRUE)}
}
\value{
vector of occupancy probabilities or posterior predictive samples
}
\description{
Forward simulation for a dynamic model
}
