% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_positions}
\alias{get_positions}
\title{Get a matrix/array of row numbers from flocker_data$data in the shape of obs}
\usage{
get_positions(data_object, unit_level = FALSE)
}
\arguments{
\item{data_object}{a flocker_fit object or a flocker_data object}

\item{unit_level}{logical. If `FALSE`, returns values associated with each visit
in the shape of obs, with NAs for visits that did not occur.
If `TRUE`, returns values associated with each unit in the shape of 
the slice of obs corresponding to the first visit. This is relevant in
multiseason models, where it is possible to have units (i.e. timesteps) 
that are part of the timeseries and have linear predictors for colonization 
etc, but that received no visits. These units are dropped from the 
formatted data if they are trailing (at the end of the series) but must be 
included otherwise. If `TRUE`, the return will have NAs for the trailing 
units that are dropped, but will have row-numbers representing the rows 
giving the unit covariates for the never-visited units if the timesteps are 
not trailing.}
}
\value{
a matrix/array of the same shape as the observations passed to 
  `make_flocker_data` (for an augmented model, in the shape of the augmented
  data created after augmentation) where each element gives the row number in 
  flocker_data$data where the corresponding element resides. If the data
  were formatted for a data-augmented model, the returned array will have
  extra slices for all of the augmented species.
}
\description{
Get a matrix/array of row numbers from flocker_data$data in the shape of obs
}
