% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_positions_V}
\alias{get_positions_V}
\title{Get matrix positions corresponding to each row of data in rep-varying 
flocker_fit}
\usage{
get_positions_V(flocker_fit)
}
\arguments{
\item{flocker_fit}{a visit-varying `flocker_fit` object}
}
\value{
an n_row x 2 matrix, where each row contains the indices of the 
    corresponding sampling event in the observation dataframe
}
\description{
Get matrix positions corresponding to each row of data in rep-varying 
flocker_fit
}
