% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Z.R
\name{get_Z_augmented}
\alias{get_Z_augmented}
\title{get Z matrix for data-augmented model}
\usage{
get_Z_augmented(lps, sample, history_condition, obs = NULL)
}
\arguments{
\item{lps}{the linear predictors from the model}

\item{sample}{logical: return fitted probabilities or bernoulli samples}

\item{history_condition}{logical: condition on the observed history?}

\item{obs}{if history_condition is true, the observed histories}
}
\value{
an array of fitted Z probabilities or sampled Z values. Rows are
  units and columns are posterior iterations.
}
\description{
get Z matrix for data-augmented model
}
