% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Z.R
\name{get_Z_dynamic}
\alias{get_Z_dynamic}
\title{get Z matrix for dynamic model}
\usage{
get_Z_dynamic(init, colo, ex, history_condition, obs = NULL, det = NULL)
}
\arguments{
\item{init}{matrix of initial occupancy probabilities (rows are sites and 
columns are draws)}

\item{colo}{array of colonization probabilities (rows are sites, columns are
timesteps, slices are draws)}

\item{ex}{array of extinction probabilities (rows are sites, columns are
timesteps, slices are draws)}

\item{history_condition}{should we condition on the observed history}

\item{obs}{Array of observations. Ignored (and defaults to NULL) if history_condition
is FALSE. Rows are sites, columns are visits, and slices are timesteps.}

\item{det}{Array of detection probabilities. Ignored (and defaults to NULL) if
history_condition is FALSE. Rows are sites, columns are visits, slices
are timesteps, and slice_2s are draws.}

\item{sample}{Should the return be posterior probabilities of occupancy (FALSE),
valid posterior predictive samples (TRUE).}
}
\value{
an array of occupancy probabilities or Z samples. Rows are sites, 
  columns are timesteps, and slices are draws.
}
\description{
get Z matrix for dynamic model
}
\details{
History-conditioning is via the forward-backward algorithm (forward
 filter backward sample when sample is TRUE)
}
