% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Z.R
\name{get_Z_single_C}
\alias{get_Z_single_C}
\title{get Z matrix for single-season rep-constant model}
\usage{
get_Z_single_C(lps, sample, history_condition, obs = NULL)
}
\arguments{
\item{lps}{the linear predictors from the model}

\item{sample}{logical: return fitted probabilities or bernoulli samples}

\item{history_condition}{logical: condition on the observed history?}

\item{obs}{if history_condition is true, the observed histories}
}
\value{
a matrix of fitted Z probabilities or sampled Z values. Rows are
  units and columns are posterior iterations.
}
\description{
get Z matrix for single-season rep-constant model
}
