% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_flocker_data}
\alias{is_flocker_data}
\title{check that object is a flocker data object}
\usage{
is_flocker_data(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical; TRUE if x is a flocker data object
}
\description{
check that object is a flocker data object
}
