% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_flocker_fit}
\alias{is_flocker_fit}
\title{Test whether an object is of class flocker_fit}
\usage{
is_flocker_fit(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
logical
}
\description{
Test whether an object is of class flocker_fit
}
