% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_flocker_formula}
\alias{is_flocker_formula}
\title{check that object is a valid flocker formula}
\usage{
is_flocker_formula(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical; TRUE if x is a valid flocker formula
}
\description{
check that object is a valid flocker formula
}
