% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_formula}
\alias{is_formula}
\title{check that object is a formula}
\usage{
is_formula(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical; TRUE if x is a formula
}
\description{
check that object is a formula
}
