% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_named_list}
\alias{is_named_list}
\title{check that object is a named list with no duplicate names}
\usage{
is_named_list(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical; TRUE if x is a named list with no duplicate names
}
\description{
check that object is a named list with no duplicate names
}
