% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_one_logical}
\alias{is_one_logical}
\title{check if an object is a single logical value}
\usage{
is_one_logical(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical; TRUE if x is a single logical value
}
\description{
check if an object is a single logical value
}
