% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_one_pos_int}
\alias{is_one_pos_int}
\title{check if an object is a single integer (or integerish) > m}
\usage{
is_one_pos_int(x, m = 0)
}
\arguments{
\item{x}{object to test}

\item{minimum}{value for x}
}
\value{
logical; TURE if x is a single positive integer
}
\description{
check if an object is a single integer (or integerish) > m
}
