% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_inv_logit}
\alias{log_inv_logit}
\title{Numerically stable log inverse logit}
\usage{
log_inv_logit(x)
}
\arguments{
\item{x}{real number or vector of reals}
}
\value{
the logarithm of the inverse logit of x
}
\description{
Numerically stable log inverse logit
}
\examples{
log_inv_logit(0)
}
