% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_occupancy.R
\name{log_lik_dynamic}
\alias{log_lik_dynamic}
\title{log_lik_dynamic}
\usage{
log_lik_dynamic(init, colo, ex, obs, det)
}
\arguments{
\item{init}{matrix of initial occupancy probabilities (rows are series and 
columns are draws)}

\item{colo}{array of colonization probabilities (rows are series, columns are
timesteps, slices are draws)}

\item{ex}{array of extinction probabilities (rows are series, columns are
timesteps, slices are draws)}

\item{obs}{Array of observations. Rows are series, columns are visits, and 
slices are timesteps.}

\item{det}{Array of detection probabilities. Rows are series, columns are 
visits, slices are timesteps, and slice_2s are draws.}
}
\value{
A log-likelihood matrix. Rows are sites, 
  columns are draws.
}
\description{
log_lik_dynamic
}
