% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_occupancy.R
\name{log_lik_occupancy_C}
\alias{log_lik_occupancy_C}
\title{A log-likelihood function for the visit-constant occupancy model, sufficient for
\code{brms::loo(vc_fit)} to work.}
\usage{
log_lik_occupancy_C(i, prep)
}
\arguments{
\item{i}{Posterior iteration}

\item{prep}{Output of \code{brms::prepare_predictions}. See brms custom families
vignette at 
https://cran.r-project.org/web/packages/brms/vignettes/brms_customfamilies.html}
}
\value{
The log-likelihood for observation i
}
\description{
A log-likelihood function for the visit-constant occupancy model, sufficient for
\code{brms::loo(vc_fit)} to work.
}
