% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_V.R
\name{log_lik_V_iter}
\alias{log_lik_V_iter}
\title{Compute sitewise log likelihood for a single posterior iteration for a 
flocker_fit object with visit-variable covariates}
\usage{
log_lik_V_iter(n_site, visit_index_matrix, resp, Q, lpo_iter, lpd_iter)
}
\arguments{
\item{n_site}{number of sites or species-sites in the data}

\item{visit_index_matrix}{a matrix or dataframe giving the first n_site rows
of the matrix of visit indices in the data (subsequent rows are redundant)}

\item{resp}{a vector giving the integer response data}

\item{Q}{a vector giving the first n_site elements of Q (whether there is at
least one detection at the site)}

\item{lpo_iter}{a vector giving the first n_site elements of the logit-scale 
linear predictor for occupancy for the iteration of interest}

\item{lpd_iter}{a vector giving the first n_site elements of the logit-scale 
linear predictor for detection for the iteration of interest}
}
\value{
a vector of sitewise log-likelihood values for the iteration of interest
}
\description{
Compute sitewise log likelihood for a single posterior iteration for a 
flocker_fit object with visit-variable covariates
}
