% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_V.R
\name{log_lik_V_site}
\alias{log_lik_V_site}
\title{Compute the likelihood for a single site in a single posterior iteration for a
flocker_fit object with visit-variable covariates}
\usage{
log_lik_V_site(resp_i, Q_i, lpo_i, lpd_i)
}
\arguments{
\item{resp_i}{the response vector (detection/non-detection) at the site}

\item{Q_i}{whether there is at least one detection, equivalent to
(\code{as.integer(sum(resp_i) > 1)})}

\item{lpo_i}{the logit-scale linear predictor for the site and iteration}

\item{lpd_i}{a vector (or scalar if the site has just one visit) giving the 
logit-scale linear predictor for the site and iteration}
}
\description{
Compute the likelihood for a single site in a single posterior iteration for a
flocker_fit object with visit-variable covariates
}
