% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo_compare_flock}
\alias{loo_compare_flock}
\title{LOO comparisons for flocker models.}
\usage{
loo_compare_flock(x, model_names = NULL)
}
\arguments{
\item{x}{a list of flocker_fit objects.}

\item{model_names}{An optional vector of names for the models.}
}
\description{
LOO comparisons for flocker models.
}
