% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo_flock_onefit}
\alias{loo_flock_onefit}
\title{Compute loo for a single flocker_fit object}
\usage{
loo_flock_onefit(x)
}
\arguments{
\item{x}{a flocker_fit object}
}
\value{
a loo object
}
\description{
Compute loo for a single flocker_fit object
}
