% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_colex_likelihoods}
\alias{make_colex_likelihoods}
\title{Create Stan code for transition log-likelihoods in colex model.
These likelihoods ignore disallowed possibilities (detections on unoccupied
units); such possibilities are excluded via the control flow in the 
implementation of the forward algorithm.}
\usage{
make_colex_likelihoods()
}
\value{
Character string of Stan code corresponding to likelihoods for the 
 possible transitions in a colex model
}
\description{
Create Stan code for transition log-likelihoods in colex model.
These likelihoods ignore disallowed possibilities (detections on unoccupied
units); such possibilities are excluded via the control flow in the 
implementation of the forward algorithm.
}
