% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_emission_0_fp}
\alias{make_emission_0_fp}
\title{Create Stan code for the emission log-likelihood given that the true state 
  equals zero in a fp model.}
\usage{
make_emission_0_fp()
}
\value{
Character string of Stan code corresponding to emission_0_fp
}
\description{
Create Stan code for the emission log-likelihood given that the true state 
  equals zero in a fp model.
}
