% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_emission_1_C}
\alias{make_emission_1_C}
\title{Create Stan code for the emission log-likelihood given that the true state 
  equals one in a rep-constant model.}
\usage{
make_emission_1_C()
}
\value{
Character string of Stan code corresponding to emission_1
}
\description{
Create Stan code for the emission log-likelihood given that the true state 
  equals one in a rep-constant model.
}
