% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_emission_1_fp_legacy}
\alias{make_emission_1_fp_legacy}
\title{Create legacy Stan code for the emission log-likelihood given that the true state 
  equals one in a fp model.}
\usage{
make_emission_1_fp_legacy()
}
\value{
Character string of Stan code corresponding to emission_1_fp
}
\description{
Create legacy Stan code for the emission log-likelihood given that the true state 
  equals one in a fp model.
}
\details{
this legacy function is deprecated; it has been verified to be
  numberically equivalent to the much simpler and more efficient 
  `make_emission_1_fp`. I'm keeping it around because I can imagine a
  possibility that it could be useful to directly marginalize over all the 
  possible histories in a model where detections conditional on occupancy
  are expected to be autocorrelated. As of renaming and deprecation of this, 
  function in April 2023, there are no plans to enable such a model in 
  flocker.
}
