% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_forward_colex}
\alias{make_forward_colex}
\title{Create Stan code for forward_colex function, an implementation of the 
 forward algorithm for use in colex models.}
\usage{
make_forward_colex()
}
\value{
Character string of Stan code corresponding to forward_colex
}
\description{
Create Stan code for forward_colex function, an implementation of the 
 forward algorithm for use in colex models.
}
