% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_forward_colex_fp}
\alias{make_forward_colex_fp}
\title{Create Stan code for forward_colex_fp function, an implementation of the 
 forward algorithm for use in colex_fp models.}
\usage{
make_forward_colex_fp()
}
\value{
Character string of Stan code corresponding to forward_colex_fp
}
\description{
Create Stan code for forward_colex_fp function, an implementation of the 
 forward algorithm for use in colex_fp models.
}
