% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_occupancy_multi_autologistic_eq_lpmf}
\alias{make_occupancy_multi_autologistic_eq_lpmf}
\title{Create Stan code for likelihood function occupancy_multi_autologistic_eq_lpmf.}
\usage{
make_occupancy_multi_autologistic_eq_lpmf(max_rep, max_year)
}
\arguments{
\item{max_rep}{Literal integer maximum number of repeated sampling events at 
any unit.}

\item{max_year}{Literal integer maximum number of years (or seasons) visited
in any series}
}
\value{
Character string of Stan code corresponding to occupancy_multi_autologistic_eq_lpmf
}
\description{
Create Stan code for likelihood function occupancy_multi_autologistic_eq_lpmf.
}
