% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_occupancy_multi_autologistic_lpmf}
\alias{make_occupancy_multi_autologistic_lpmf}
\title{Create Stan code for likelihood function occupancy_multi_autologistic_lpmf.}
\usage{
make_occupancy_multi_autologistic_lpmf(max_rep, max_year)
}
\arguments{
\item{max_rep}{Literal integer maximum number of repeated sampling events at 
any unit.}
}
\value{
Character string of Stan code corresponding to occupancy_multi_autologistic_lpmf
}
\description{
Create Stan code for likelihood function occupancy_multi_autologistic_lpmf.
}
