% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_occupancy_multi_colex_fp_lpdf}
\alias{make_occupancy_multi_colex_fp_lpdf}
\title{Create Stan code for likelihood function occupancy_multi_colex_fp_lpdf.}
\usage{
make_occupancy_multi_colex_fp_lpdf(max_rep, max_year)
}
\arguments{
\item{max_rep}{Literal integer maximum number of repeated sampling events at 
any unit.}
}
\value{
Character string of Stan code corresponding to occupancy_multi_colex_fp_lpdf
}
\description{
Create Stan code for likelihood function occupancy_multi_colex_fp_lpdf.
}
