% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_occupancy_single_C_lpmf}
\alias{make_occupancy_single_C_lpmf}
\title{Create Stan code for likelihood function occupancy_single_C_lpmf.
The purpose of defining this custom family, rather than using brms's zero-inflated 
binomial, is to ensure that the occupancy parameters are interpretable with
values of 1 in the marginalized state reflecting occupancy rather than non-
occupancy.}
\usage{
make_occupancy_single_C_lpmf()
}
\value{
Character string of Stan code corresponding to occupancy_single_C_lpmf
}
\description{
Create Stan code for likelihood function occupancy_single_C_lpmf.
The purpose of defining this custom family, rather than using brms's zero-inflated 
binomial, is to ensure that the occupancy parameters are interpretable with
values of 1 in the marginalized state reflecting occupancy rather than non-
occupancy.
}
