% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_occupancy_single_fp_lpdf}
\alias{make_occupancy_single_fp_lpdf}
\title{Create Stan code for likelihood function occupancy_single_fp_lpmf}
\usage{
make_occupancy_single_fp_lpdf(max_rep)
}
\arguments{
\item{max_rep}{Literal integer maximum number of repeated sampling events at 
any unit.}
}
\value{
Character string of Stan code corresponding to occupancy_single_fp_lpmf
}
\description{
Create Stan code for likelihood function occupancy_single_fp_lpmf
}
