% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_occupancy_vc_lpmf}
\alias{make_occupancy_vc_lpmf}
\title{Create Stan code occupancy_vc_lpmf with visit-constant covariates
Primarily for internal use in \code{flock()}.}
\usage{
make_occupancy_vc_lpmf()
}
\value{
Character string of Stan code corresponding to occupancy_vc_lpmf
}
\description{
Create Stan code occupancy_vc_lpmf with visit-constant covariates
Primarily for internal use in \code{flock()}.
}
