% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_occupancy_vv_lpmf}
\alias{make_occupancy_vv_lpmf}
\title{Create Stan code for occupancy_vv_lpmf with visit-varying covariates
Primarily for internal use in \code{flock()}.}
\usage{
make_occupancy_vv_lpmf(max_visit)
}
\arguments{
\item{max_visit}{Literal integer maximum number of visits to any site.}
}
\value{
Character string of Stan code corresponding to occupancy_vv_lpmf
}
\description{
Create Stan code for occupancy_vv_lpmf with visit-varying covariates
Primarily for internal use in \code{flock()}.
}
