% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mfd_single}
\alias{mfd_single}
\title{Data for example single-season flocker model}
\format{
## `mfd_single`
A flocker_fit and brmsfit object
}
\source{
data-raw/example_flocker_model.R
}
\usage{
mfd_single
}
\description{
A flocker_data object for a single-season occupancy model
}
\keyword{datasets}
