% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multi_init_types}
\alias{multi_init_types}
\title{possible values for the `multi_init` parameter if not NULL}
\usage{
multi_init_types()
}
\value{
character vector of options
}
\description{
possible values for the `multi_init` parameter if not NULL
}
