% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multi_types}
\alias{multi_types}
\title{flocker_data_output_types valid for use with colex models}
\usage{
multi_types()
}
\value{
character vector of data output types
}
\description{
flocker_data_output_types valid for use with colex models
}
