% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_array}
\alias{new_array}
\title{create a new array with the dimensions of an old array}
\usage{
new_array(m, data = NA)
}
\arguments{
\item{m}{old array}

\item{data}{passed to the data argument of array()}
}
\value{
new array
}
\description{
create a new array with the dimensions of an old array
}
