% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_matrix}
\alias{new_matrix}
\title{create a new matrix with the dimensions of an old matrix}
\usage{
new_matrix(m, data = NA, byrow = FALSE)
}
\arguments{
\item{m}{old matrix}

\item{data}{passed to the data argument of matrix()}

\item{byrow}{passed to the byrow argument of matrix()}
}
\value{
new matrix
}
\description{
create a new matrix with the dimensions of an old matrix
}
