% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_families.R
\name{occupancy_augmented}
\alias{occupancy_augmented}
\title{Define the rep-varying augmented occupancy family}
\usage{
occupancy_augmented(max_rep)
}
\value{
a "customfamily" "brmsfamily" object from brms
}
\description{
Define the rep-varying augmented occupancy family
}
