% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_occupancy.R
\name{occupancy_C_lpmf}
\alias{occupancy_C_lpmf}
\title{An R implementation of the rep constant lpmf without the binomial coefficient.}
\usage{
occupancy_C_lpmf(y, mu, occ, trials)
}
\arguments{
\item{y}{number of detections}

\item{mu}{logit-scale detection probability}

\item{occ}{logit-scale occupancy probability}

\item{trials}{number of reps}
}
\value{
The log-likelihood
}
\description{
An R implementation of the rep constant lpmf without the binomial coefficient.
}
\details{
By omitting the binomial coefficient, the log-likelihood is directly
  comparable to log-likelihoods from rep-varying models.
}
