% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_families.R
\name{occupancy_multi_autologistic_eq}
\alias{occupancy_multi_autologistic_eq}
\title{Define the autologistic_eq family}
\usage{
occupancy_multi_autologistic_eq(max_year, max_rep)
}
\arguments{
\item{max_year}{the maximum number of seasons at a colex unit}

\item{max_rep}{the maximum number of repeat sampling events at a closure unit}
}
\value{
a "customfamily" "brmsfamily" object from brms
}
\description{
Define the autologistic_eq family
}
