% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_families.R
\name{occupancy_single_C}
\alias{occupancy_single_C}
\title{Define the rep-constant occupancy family}
\usage{
occupancy_single_C()
}
\value{
a "customfamily" "brmsfamily" object from brms
}
\description{
Define the rep-constant occupancy family
}
