% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_families.R
\name{occupancy_single_fp}
\alias{occupancy_single_fp}
\title{Define the single-fp family}
\usage{
occupancy_single_fp(max_rep)
}
\arguments{
\item{max_rep}{the maximum number of repeat sampling events at a unit}
}
\value{
a "customfamily" "brmsfamily" object from brms
}
\description{
Define the single-fp family
}
