% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_families.R
\name{occupancy_single_threaded}
\alias{occupancy_single_threaded}
\title{Define the rep-varying occupancy family with threading}
\usage{
occupancy_single_threaded(max_rep)
}
\arguments{
\item{max_rep}{the maximum number of repeat sampling events at a unit}
}
\value{
a "customfamily" "brmsfamily" object from brms
}
\description{
Define the rep-varying occupancy family with threading
}
