% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_families.R
\name{occupancy_V}
\alias{occupancy_V}
\title{Define the rep-varying occupancy family
Primarily for internal use in \code{flock()}.}
\usage{
occupancy_V(max_rep)
}
\arguments{
\item{max_rep}{the maximum number of repeat sampling events at a unit}
}
\value{
a "customfamily" "brmsfamily" object from brms
}
\description{
Define the rep-varying occupancy family
Primarily for internal use in \code{flock()}.
}
