% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_families.R
\name{occupancy_vc}
\alias{occupancy_vc}
\title{Define the occupancy model family with visit-constant covariates
Primarily for internal use in \code{flock()}.}
\usage{
occupancy_vc()
}
\value{
a "customfamily" "brmsfamily" object from brms
}
\description{
Define the occupancy model family with visit-constant covariates
Primarily for internal use in \code{flock()}.
}
