% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{params_by_type}
\alias{params_by_type}
\title{A list of distributional parameters by model type}
\format{
An object of class \code{list} of length 10.
}
\usage{
params_by_type
}
\value{
A list of distributional parameters by model type
}
\description{
A list of distributional parameters by model type
}
\keyword{datasets}
