% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_C-prediction-epred.R
\name{posterior_predict_occupancy_single_C}
\alias{posterior_predict_occupancy_single_C}
\title{A posterior predict function for the rep-constant occupancy model, sufficient 
for \code{brms::predict.brmsfit()} to work.}
\usage{
posterior_predict_occupancy_single_C(i, prep, ...)
}
\arguments{
\item{i}{Observation id}

\item{prep}{Output of \code{brms::prepare_predictions}. See brms custom 
families vignette at 
https://cran.r-project.org/web/packages/brms/vignettes/brms_customfamilies.html}
}
\value{
Posterior predictions
}
\description{
A posterior predict function for the rep-constant occupancy model, sufficient 
for \code{brms::predict.brmsfit()} to work.
}
