% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.flocker_fit.R
\name{print.flocker_fit}
\alias{print.flocker_fit}
\alias{print.flocker_summary}
\title{Print a summary for a fitted model represented by a \code{flocker_fit} object.}
\usage{
\method{print}{flocker_fit}(flocker_fit, ...)
}
\arguments{
\item{flocker_fit}{An object of class \code{flocker_fit}.}

\item{...}{Other potential arguments passed to method \code{summary} of \code{flocker_fit}.}
}
\description{
Print a summary for a fitted model represented by a \code{flocker_fit} object.
}
\seealso{
\code{\link{summary.flocker_fit}}
}
