% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_rownames}
\alias{remove_rownames}
\title{remove rownames}
\usage{
remove_rownames(m)
}
\arguments{
\item{m}{object whose rownames to remove}
}
\value{
m with rownames removed
}
\description{
remove rownames
}
