% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep_constant_types}
\alias{rep_constant_types}
\title{flocker_data_output_types that yield rep-constant models}
\usage{
rep_constant_types()
}
\value{
character vector of data output types
}
\description{
flocker_data_output_types that yield rep-constant models
}
