% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shared_elements}
\alias{shared_elements}
\title{get shared elements between two vectors}
\usage{
shared_elements(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
all unique elements that occur at least once in both x and y
}
\description{
get shared elements between two vectors
}
