% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stack_matrix}
\alias{stack_matrix}
\title{rbind a matrix or dataframe to itself n times}
\usage{
stack_matrix(m, n)
}
\arguments{
\item{m}{a matrix or dataframe}

\item{n}{number of times to rbind}
}
\value{
matrix or dataframe
}
\description{
rbind a matrix or dataframe to itself n times
}
