% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_flocker.R
\name{summarise_fun}
\alias{summarise_fun}
\title{function to summarize matrix of linear predictors to its mean and CI}
\usage{
summarise_fun(x, CI)
}
\arguments{
\item{x}{linpreds}
}
\value{
summary
}
\description{
function to summarize matrix of linear predictors to its mean and CI
}
