% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{threading_types}
\alias{threading_types}
\title{flocker_data_output_types valid for use with threading}
\usage{
threading_types()
}
\value{
character vector of data output types
}
\description{
flocker_data_output_types valid for use with threading
}
