% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{type_flocker_fit}
\alias{type_flocker_fit}
\title{Extract lik_type from object of class flocker_fit}
\usage{
type_flocker_fit(x)
}
\arguments{
\item{x}{flocker_fit object}
}
\value{
string giving model type
}
\description{
Extract lik_type from object of class flocker_fit
}
