% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_param_combos_augmented}
\alias{validate_param_combos_augmented}
\title{Check validity of params passed to `flock` if `type` is `augmented`}
\usage{
validate_param_combos_augmented(
  f_occ,
  f_det,
  flocker_data,
  multiseason,
  f_col,
  f_ex,
  colex_init,
  f_auto,
  augmented,
  fp,
  threads
)
}
\value{
silent if parameters are valid
}
\description{
Check validity of params passed to `flock` if `type` is `augmented`
}
