% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Z_from_emission}
\alias{Z_from_emission}
\title{get Z given emission likelihoods and state probability}
\usage{
Z_from_emission(el0, el1, psi_unconditional)
}
\arguments{
\item{el0}{emission likelihood given state 0}

\item{el1}{emission likelihood given state 1}

\item{psi_unconditional}{occupancy probability not conditioned on observation}
}
\value{
occupancy probability conditioned on observation
}
\description{
get Z given emission likelihoods and state probability
}
