% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_V.R
\name{calc_log_lik_partial}
\alias{calc_log_lik_partial}
\title{Compute the part of the log-likelihood relating to visits. To be used 
internally in log_lik_V(). Missed visits are returned as 0s}
\usage{
calc_log_lik_partial(resp, Q, lpd)
}
\arguments{
\item{resp}{the response vector (detection/non-detection) at the site. Missing 
visits are represented as -99}

\item{Q}{whether there is at least one detection at a species:point combination}

\item{lpd}{the logit-scale linear predictor}
}
\description{
Compute the part of the log-likelihood relating to visits. To be used 
internally in log_lik_V(). Missed visits are returned as 0s
}
