% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emission_likelihood}
\alias{emission_likelihood}
\title{Get emission likelihoods given observations and detection probabilities}
\usage{
emission_likelihood(state, obs, det)
}
\arguments{
\item{state}{Compute the emission likelihood for absence (0) or presence (1)}

\item{obs}{The observation history for the unit}

\item{det}{The detection probabilities at the unit}
}
\value{
the emission likelihood
}
\description{
Get emission likelihoods given observations and detection probabilities
}
