% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_flocker_data.R
\name{example_flocker_data}
\alias{example_flocker_data}
\title{Create example data for use with \code{make_flocker_data()} and downstream functions.}
\usage{
example_flocker_data(visit_constant = FALSE, seed = 123)
}
\arguments{
\item{visit_constant}{logical: return data with constant detection covariates
within sites (TRUE) or variable detection covariates (FALSE)}

\item{seed}{random seed}
}
\value{
A three element named list with the observation matrix ($obs), the
site covariate dataframe ($site_covs), and the visit covariate list
($visit_covs). If visit_constant is TRUE, then $visit_covs will be NULL.
}
\description{
Create example data for use with \code{make_flocker_data()} and downstream functions.
}
