% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_flocker_model_single}
\alias{example_flocker_model_single}
\title{Example single-season flocker model}
\format{
## `example_flocker_model_single`
A flocker_fit and brmsfit object
}
\source{
data-raw/example_flocker_model.R
}
\usage{
example_flocker_model_single
}
\description{
A fitted single-season occupancy model from flocker
}
\keyword{datasets}
