% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flock.R
\name{flock}
\alias{flock}
\title{Fit an occupancy model}
\usage{
flock(f_occ, f_det, flocker_data, data2 = NULL, rep_constant = FALSE, ...)
}
\arguments{
\item{f_occ}{A brms-type model formula for occupancy. Must begin with "~".}

\item{f_det}{A brms-type model formula for detection. Must begin with "~".}

\item{flocker_data}{data, generally the output of \code{make_flocker_data()}.}

\item{data2}{additional data (e.g. a covariance matrix for a phylogenetic effect)}

\item{rep_constant}{A logical indicator. Are detection probabilities constant 
across repeated sampling events within closure units?}

\item{...}{additional arguments passed to \code{brms::brm()}}
}
\value{
a \code{brmsfit} containing the fitted occupancy model.
}
\description{
Fit an occupancy model
}
\examples{
\dontrun{
example_data <- example_flocker_data()
fd <- make_flocker_data(example_data$obs, example_data$unit_covs, example_data$event_covs)
flock(f_occ = ~ uc1 + s(uc2) + (1|grp),
          f_det = ~ uc1 + ec1 + s(ec2) + (1|grp),
          flocker_data = fd,
          refresh = 50, chains = 1, warmup = 5, iter = 200,
          adapt_engaged = F, step_size = .05, max_treedepth = 5, seed = 123)
}
}
