% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flock.R
\name{flocker_fit_code_util}
\alias{flocker_fit_code_util}
\title{Either fit a flocker model or output stancode}
\usage{
flocker_fit_code_util(stancode, f_use, data, data2, family, stanvars, ...)
}
\arguments{
\item{stancode}{logical; if TRUE return stan code, if false fit model}

\item{f_use}{brms formula}

\item{data}{brms data}

\item{data2}{brms data2}

\item{family}{brms family}

\item{stanvars}{brms stanvars}

\item{...}{brms ...}
}
\value{
output of brms::make_stancode or brms::brm
}
\description{
Either fit a flocker model or output stancode
}
