% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flocker_make_standata.R
\name{flocker_make_standata}
\alias{flocker_make_standata}
\title{Make Stan data for a flocker model with visit-specific detection covariates.
Primarily for internal use in \code{flocker()}.}
\usage{
flocker_make_standata(model_formula, flocker_data, data2)
}
\arguments{
\item{model_formula}{The model formula (occupancy + detection)}

\item{flocker_data}{data, generally the output of \code{make_flocker_data()}.}

\item{data2}{additional data (e.g. a covariance matrix for a phylogenetic effect)}
}
\value{
A standata list.
}
\description{
Make Stan data for a flocker model with visit-specific detection covariates.
Primarily for internal use in \code{flocker()}.
}
