% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Z.R
\name{forward_algorithm}
\alias{forward_algorithm}
\title{Forward algorithm}
\usage{
forward_algorithm(e0, e1, init, colo, ex)
}
\arguments{
\item{e0}{the emission probabilities given non-occupancy}

\item{e1}{the emission probabilities given occupancy}

\item{init}{the initial state probabilities}

\item{colo}{a vector giving the colonization probs}

\item{ex}{a vector giving the extinction probs}
}
\value{
an n x 2 matrix giving the forward probabilities associated with 
  non-occupancy (first column) and occupancy (second column)
}
\description{
Forward algorithm
}
