% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Z.R
\name{get_Z}
\alias{get_Z}
\title{Get posterior distribution of Z matrix}
\usage{
get_Z(
  flocker_fit,
  n_iter = NULL,
  history_condition = TRUE,
  new_data = NULL,
  sample_new_levels = "uncertainty"
)
}
\arguments{
\item{flocker_fit}{A flocker_fit object}

\item{n_iter}{The number of posterior iterations desired. If `NULL`, use
all available posterior iterations.}

\item{history_condition}{Should the posterior distribution for Z directly 
condition on the observed detection history (`TRUE`) or not (`FALSE`)?
For example, at sites with at least one detection, the true occupancy 
state conditioned on the history is one with absolute certainty. Without 
directly conditioning on the history, the occupancy state is controlled 
by the posterior distribution for the occupancy probability psi.}

\item{new_data}{Optional new data at which to predict the Z matrix. Generally
the output of `make_flocker_data`. Cannot be used if `history_condition` is
set to `TRUE`.}

\item{sample_new_levels}{If new_data is provided and contains random effect
levels not present in the original data, how should predictions be
handled? Passed directly to brms::prepare_predictions, which see.}
}
\value{
The posterior Z matrix. Rows are iterations and columns
    are closure-units. Values are samples from the posterior distribution
    of occupancy probability.
}
\description{
Get posterior distribution of Z matrix
}
