% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Z.R
\name{get_Z}
\alias{get_Z}
\title{Get posterior distribution of Z matrix}
\usage{
get_Z(flocker_fit, n_iter = NULL)
}
\arguments{
\item{flocker_fit}{A flocker_fit object}

\item{n_iter}{The number of posterior iterations desired}
}
\value{
The posterior Z matrix. Rows are iterations and columns
    are closure-units. Values are samples from the posterior distribution
    of occupancy probability.
}
\description{
Get posterior distribution of Z matrix
}
