% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log1m_inv_logit}
\alias{log1m_inv_logit}
\title{Numerically stable log one-minus inverse logit}
\usage{
log1m_inv_logit(x)
}
\arguments{
\item{x}{real number or vector of reals}
}
\value{
the logarithm of one minus the inverse logit of x
}
\description{
Numerically stable log one-minus inverse logit
}
\examples{
log1m_inv_logit(0)
}
