% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_occupancy.R
\name{log_lik_flocker}
\alias{log_lik_flocker}
\title{Compute pointwise log-likelihood matrix for a flocker_fit object}
\usage{
log_lik_flocker(
  flocker_fit,
  new_data = NULL,
  allow_new_levels = FALSE,
  sample_new_levels = "uncertainty",
  draw_ids = NULL
)
}
\arguments{
\item{flocker_fit}{A flocker_fit object}

\item{new_data}{optional new data at which to compute log likelihood}

\item{allow_new_levels}{allow new levels for random effect terms in 
'new_data'? Will error if set to 'FALSE' and new levels are provided in 
'new_data'.}

\item{sample_new_levels}{If new_data is provided and contains random effect 
levels not present in the original data, how should predictions be 
handled? See '?brms::prepare_predictions' for options.}

\item{draw_ids}{the draw ids to compute log-likelihoods for. Defaults to 
using the full posterior.}
}
\value{
A posterior log-likelihood matrix, where iterations are rows and
  likelihood units are columns. For two-level models, columns are named with
  the level-two group names.
}
\description{
Compute pointwise log-likelihood matrix for a flocker_fit object
}
\details{
In one-level single-season models, columns are closure units (e.g.
  points or species-points; suitable for leave-one-unit-out CV). In
  multiseason models, columns are series (i.e. points or species-points;
  suitable for leave-one-series-out CV). In `twolevel_single` models,
  columns are level-two groups. In augmented models, columns are species
  (suitable for leave-one-species-out CV).
}
\examples{
\dontrun{
log_lik_flocker(example_flocker_model_single)
}
}
