% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_occupancy.R
\name{log_lik_flocker}
\alias{log_lik_flocker}
\title{Compute unit-wise or series-wise log-likelihood matrix for a flocker_fit object}
\usage{
log_lik_flocker(flocker_fit, draw_ids = NULL)
}
\arguments{
\item{flocker_fit}{A flocker_fit object}

\item{draw_ids}{the draw ids to compute log-likelihoods for. Defaults to using the full posterior.}
}
\value{
A unit-wise or series-wise posterior log-likelihood matrix, where iterations are rows and 
   units/series are columns
}
\description{
Compute unit-wise or series-wise log-likelihood matrix for a flocker_fit object
}
\examples{
\dontrun{
log_lik_flocker(example_flocker_model_single)
}
}
