% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_occupancy.R
\name{log_lik_V}
\alias{log_lik_V}
\title{Compute unitwise log-likelihood matrix for a flocker_fit object with 
visit-variable covariates}
\usage{
log_lik_V(flocker_fit_V)
}
\arguments{
\item{flocker_fit_V}{A flocker_fit with visit-variable covariates}
}
\value{
A unitwise posterior log-likelihood matrix
}
\description{
Compute unitwise log-likelihood matrix for a flocker_fit object with 
visit-variable covariates
}
