% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo_compare_flocker}
\alias{loo_compare_flocker}
\title{LOO comparisons for flocker models.}
\usage{
loo_compare_flocker(model_list, model_names = NULL, thin = NULL)
}
\arguments{
\item{model_list}{a list of flocker_fit objects.}

\item{model_names}{An optional vector of names for the models.}

\item{thin}{specify the amount of thinning required. 1 or NULL results in no 
thinning, 2 retains every other value, 3 every third, etc.}
}
\value{
a `compare.loo` matrix
}
\description{
LOO comparisons for flocker models.
}
\examples{
\donttest{
ml <- rep(list(example_flocker_model_single), 3)
loo_compare_flocker(ml)
}
}
