% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo_flocker}
\alias{loo_flocker}
\title{Compute loo for flocker_fit objects}
\usage{
loo_flocker(x)
}
\arguments{
\item{x}{a flocker_fit object or a list of flocker_fit objects}
}
\value{
a loo object or a list of loo objects
}
\description{
Compute loo for flocker_fit objects
}
