% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flocker_data.R
\name{make_flocker_data}
\alias{make_flocker_data}
\title{Format data for \code{flocker()}.}
\usage{
make_flocker_data(obs, constant_covs = NULL, visit_covs = NULL)
}
\arguments{
\item{obs}{An I x J matrix-like object where closure is assumed across rows (i.e. 
rows are sampling sites or species-sites) and columns are visits. Allowable 
values are 1 (detection), 0 (no detection), and NA (no visit).}

\item{constant_covs}{A dataframe of covariates for each unit that are constant 
across visits.}

\item{visit_covs}{A named list of I x J matrices, each one corresponding to a covariate
that varies by visit.}
}
\value{
A flocker_data list that can be passed as data to \code{flocker()}.
}
\description{
Format data for \code{flocker()}.
}
