% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flocker_data.R
\name{make_flocker_data_twolevel_single}
\alias{make_flocker_data_twolevel_single}
\title{Format data for two-level single-season occupancy model, to be passed to
\code{flock()}.}
\usage{
make_flocker_data_twolevel_single(
  obs,
  unit_covs,
  event_covs = NULL,
  group_covs,
  top_level,
  known_present = NULL,
  quiet = FALSE,
  newdata_checks = FALSE
)
}
\arguments{
\item{obs}{If \code{type = "single"}, an I x J matrix-like object where 
closure is assumed across rows and columns are repeated sampling events. 
  If \code{type = "multi"}, an I x J x K array where rows are sites or 
species-sites, columns are repeated sampling events, and slices along the 
third dimension are seasons. Allowable values are 1 (detection), 0 (no 
detection), and NA (no sampling event).
   If \code{type = "augmented"}, an L x J x K array where rows L are sites, 
columns J are repeat sampling events, and slices K are species. 
   The data must be packed so that, for a given unit (site, site-species, 
site-timestep, site-species-timestep) all realized visits come before any 
missing visits (NAs are trailing within their rows).}

\item{unit_covs}{If \code{type = "single"} a dataframe of covariates for each 
closure-unit that are constant across repeated sampling events within units.
  If \code{type = "multi"}, a list of such dataframes, one per timestep. All 
dataframes must have identical column names and types, and all
dataframes must have I rows.
  If \code{type = "augmented"}, a dataframe of covariates for each site that
are constant across repeated sampling events within sites (no dependence on
species is allowed).}

\item{event_covs}{If \code{type = "single"}, a named list of I x J matrices, 
each one corresponding to a covariate that varies across repeated sampling 
events within closure-units.
  If \code{type = "multi"}, a named list of I x J x K arrays, each one 
corresponding to a covariate that varies across repeated sampling events 
within closure-units.
  If \code{type = "augmented"}, a named list of L x J matrices, each one
corresponding to a covariate that varies across repeated sampling events
within sites (no dependence on species is allowed).}

\item{group_covs}{A dataframe of covariates for each top-level group.}

\item{top_level}{The name of a factor column in both \code{unit_covs} and
\code{group_covs} identifying the top-level group.}

\item{known_present}{Optional name of a logical column in \code{group_covs}
indicating groups known a priori to be present. Groups with detections are
always treated as known present regardless of this column.}

\item{quiet}{Hide progress bars and informational messages?}

\item{newdata_checks}{If TRUE, turn off checks that must pass in order
to use the data for model fitting, but not in other contexts (e.g. making
predictions or assessing log-likelihoods over new data).}
}
\value{
A flocker_data list that can be passed as data to \code{flock()}.
}
\description{
Format data for two-level single-season occupancy model, to be passed to
\code{flock()}.
}
