% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occupancy_lpmf.R
\name{make_occupancy_vc_lpmf}
\alias{make_occupancy_vc_lpmf}
\title{Create Stan code occupancy_vc_lpmf with visit-constant covariates
Primarily for internal use in \code{flock()}.
The purpose of defining this custom family, rather than using brms's zero-inflated 
binomial, is to ensure that the occupancy parameters are interpretable as though
values of 1 in the marginalized state reflect occupancy rather than non-
occupancy.}
\usage{
make_occupancy_vc_lpmf()
}
\value{
Character string of Stan code corresponding to occupancy_vc_lpmf
}
\description{
Create Stan code occupancy_vc_lpmf with visit-constant covariates
Primarily for internal use in \code{flock()}.
The purpose of defining this custom family, rather than using brms's zero-inflated 
binomial, is to ensure that the occupancy parameters are interpretable as though
values of 1 in the marginalized state reflect occupancy rather than non-
occupancy.
}
