% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nslice}
\alias{nslice}
\title{Get third dimension of an array}
\usage{
nslice(x)
}
\arguments{
\item{x}{array for which third dimension is desired}
}
\value{
integer size of third dimension
}
\description{
Get third dimension of an array
}
