% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_families.R
\name{occupancy_V}
\alias{occupancy_V}
\title{Define the occupancy model family with visit-varying covariates
Primarily for internal use in \code{flock()}.}
\usage{
occupancy_V(max_visit)
}
\arguments{
\item{max_visit}{the maximum number of repeat visits to a unit}
}
\value{
a "customfamily" "brmsfamily" object from brms
}
\description{
Define the occupancy model family with visit-varying covariates
Primarily for internal use in \code{flock()}.
}
