% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_flocker.R
\name{reshape_fun}
\alias{reshape_fun}
\title{function to reshape a matrix of values into a stack of arrays via get_positions}
\usage{
reshape_fun(x, gp)
}
\arguments{
\item{x}{input_matrix}

\item{gp}{output of get_positions}
}
\description{
function to reshape a matrix of values into a stack of arrays via get_positions
}
