% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_flocker_data.R
\name{sfd}
\alias{sfd}
\title{util for creating example data}
\usage{
sfd(
  n_rep,
  n_pt,
  n_sp,
  n_season,
  multiseason,
  multi_init,
  augmented,
  rep_constant,
  fp,
  params,
  covariates,
  ragged_rep,
  missing_seasons
)
}
\arguments{
\item{n_rep}{number of replicate visits to simulate per closure unit}

\item{n_pt}{number of points to simulate. The number of units for single-
season models will be `n_pt*n_sp`. The number of units for multi-season
models will be `n_pt*n_sp*n_season`.}

\item{n_sp}{number of levels to include in random effect. For compatibility
with multispecies models, this levels of this random effect define separate
closure units.}

\item{n_season}{A positive integer giving the number of seasons desired. 1
yields data for a single-season model; all other positive integers yield
multiseason models.}

\item{multiseason}{if n_season is NULL, must be NULL. Otherwise, one of
"colex" or "autologistic"}

\item{multi_init}{if n_season is NULL, must be NULL. Otherwise, one of 
"explicit" or "equilibrium"}

\item{augmented}{logical. If `TRUE` data will be formatted for an augmented
model; `fp` must be `NULL`, and `n_season` must be `1`; all never-observed 
species will be trimmed out of the data before augmenting; by default 
data will be simulated under a large random effect variance for detection 
to encourage the existence of never-observed species; and by default data 
will not include any covariate effects on occupancy.}

\item{rep_constant}{logical: create data with unit covariates only (TRUE) 
or data that includes event covariates (FALSE)}

\item{fp}{logical: if create data for fp model (TRUE) or standard model 
(FALSE). If `TRUE`, the simulation will produce detections that correspond
to true detections with probability 0.8.}

\item{params}{a named list of parameter values to use in simulation, or NULL 
to simulate values.}

\item{covariates}{a dataframe of covariate values to use in simulation, or
NULL oto simulate values.}

\item{ragged_rep}{logical: create data with variable (TRUE) or constant 
(FALSE) numbers of visits per unit.  If TRUE, approximately half of units 
will be missing approximately half of `n_rep` visits.}

\item{missing_seasons}{logical; relevant only if n_season is greater than 1. 
create data with variable (TRUE) or constant (FALSE) numbers of seasons per 
series (TRUE). If TRUE, approximately half of series will be missing their
even-numbered seasons.}
}
\value{
A three element named list with the observation matrix/array ($obs), 
  the unit covariate dataframe(s) ($unit_covs), and the event covariate list
  ($rep_covs). If rep_constant is TRUE, then $rep_covs will be NULL.
}
\description{
util for creating example data
}
