% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.flocker_fit.R
\name{summary.flocker_fit}
\alias{summary.flocker_fit}
\title{Summary method for a flocker_fit object.}
\usage{
\method{summary}{flocker_fit}(flocker_fit)
}
\arguments{
\item{flocker_fit}{A flocker_fit object produced by \code{flocker(..., visit_constant = F)}.}

\item{...}{Other potential arguments.}
}
\description{
Summary method for a flocker_fit object.
}
