% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_param_combos_multi}
\alias{validate_param_combos_multi}
\title{Check validity of params passed to `flock` if `type` is `multi`}
\usage{
validate_param_combos_multi(
  f_occ,
  f_det,
  flocker_data,
  multiseason,
  f_col,
  f_ex,
  colex_init,
  augmented,
  fp
)
}
\value{
silent if parameters are valid
}
\description{
Check validity of params passed to `flock` if `type` is `multi`
}
