% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_param_combos_single_generic}
\alias{validate_param_combos_single_generic}
\title{Check validity of some params passed to flock if `type` is `single` or `single_C`}
\usage{
validate_param_combos_single_generic(
  f_occ,
  f_det,
  flocker_data,
  multiseason,
  f_col,
  f_ex,
  multi_init,
  f_auto,
  augmented
)
}
\arguments{
\item{f_occ}{occupancy formula}

\item{f_det}{detection formula}

\item{flocker_data}{flocker_data object}

\item{multiseason}{multiseason flag}

\item{f_col}{colonization formula}

\item{f_ex}{extinction formula}

\item{multi_init}{multi_init flag}

\item{f_auto}{autologistic formula}

\item{augmented}{augmented flag}
}
\value{
silent if parameters are valid
}
\description{
Check validity of some params passed to flock if `type` is `single` or `single_C`
}
