% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_params_individually}
\alias{validate_params_individually}
\title{Check individual validity of params passed to `flock`}
\usage{
validate_params_individually(
  f_occ,
  f_det,
  flocker_data,
  multiseason,
  f_col,
  f_ex,
  colex_init,
  augmented,
  fp
)
}
\value{
silent if parameters are valid
}
\description{
Check individual validity of params passed to `flock`
}
